fix(registry/coder-labs/modules/codex): address all remaining review findings
DEREM-1: add model-reasoning-effort-standalone test (without ai_gateway)
DEREM-2: rename 'AI Bridge' to 'AI Gateway' in aibridge_config name field
DEREM-3: replace tautological var==input tftest assertions with output
  length checks that verify resource creation
DEREM-4: quote path in README coder_app example
DEREM-5: document in base_config_toml description that
  model_reasoning_effort and workdir trust are only applied in defaults
DEREM-9: extract collectScripts/runScripts/ModuleScripts into shared
  agentapi/coder-utils-test-helpers.ts, import in codex tests
DEREM-12: add negative assertions to minimal-default-config (no
  model_provider, no model_providers section, no model_reasoning_effort)
DEREM-14: already resolved by DEREM-7 base64 encoding of ARG_WORKDIR
DEREM-15: PR description issue (will update separately)
DEREM-18: debug header already uses user-facing names (fixed earlier)
DEREM-19: container is created by shared setup(); splitting would require
  a separate code path, accepted as-is since wall time is minimal
DEREM-24: add Migrating from v4 section to README
