+++ registry/coder-labs/modules/codex/main.test.ts
@@ -19,54 +19,14 @@ import {
   extractCoderEnvVars,
   writeExecutable,
 } from "../../../coder/modules/agentapi/test-util";
+import {
+  collectScripts,
+  runScripts,
+  ModuleScripts,
+} from "../../../coder/modules/agentapi/coder-utils-test-helpers";
 import path from "path";
 
-interface ModuleScripts {
-  pre_install?: string;
-  install: string;
-  post_install?: string;
-}
-
-const SCRIPT_SUFFIXES = [
-  "Pre-Install Script",
-  "Install Script",
-  "Post-Install Script",
-] as const;
-
-const collectScripts = (state: TerraformState): ModuleScripts => {
-  const byDisplayName: Record<string, string> = {};
-  for (const resource of state.resources) {
-    if (resource.type !== "coder_script") continue;
-    for (const instance of resource.instances) {
-      const attrs = instance.attributes as Record<string, unknown>;
-      const displayName = attrs.display_name as string | undefined;
-      const script = attrs.script as string | undefined;
-      if (displayName && script) {
-        byDisplayName[displayName] = script;
-      }
-    }
-  }
-  const scripts: Partial<ModuleScripts> = {};
-  for (const suffix of SCRIPT_SUFFIXES) {
-    const key = `Codex: ${suffix}`;
-    if (!(key in byDisplayName)) continue;
-    switch (suffix) {
-      case "Pre-Install Script":
-        scripts.pre_install = byDisplayName[key];
-        break;
-      case "Install Script":
-        scripts.install = byDisplayName[key];
-        break;
-      case "Post-Install Script":
-        scripts.post_install = byDisplayName[key];
-        break;
-    }
-  }
-  if (!scripts.install) {
-    throw new Error("install script not found in terraform state");
-  }
-  return scripts as ModuleScripts;
-};
+const DISPLAY_NAME_PREFIX = "Codex";
 
 let cleanupFunctions: (() => Promise<void>)[] = [];
 const registerCleanup = (cleanup: () => Promise<void>) => {
@@ -104,7 +64,7 @@ const setup = async (
     install_codex: "false",
     ...props?.moduleVariables,
   });
-  const scripts = collectScripts(state);
+  const scripts = collectScripts(state, DISPLAY_NAME_PREFIX);
   const coderEnvVars = extractCoderEnvVars(state);
 
   const id = await runContainer("codercom/enterprise-node:latest");
@@ -132,43 +92,6 @@ const setup = async (
     });
   }
   return { id, coderEnvVars, scripts };
-};
-
-const runScripts = async (
-  id: string,
-  scripts: ModuleScripts,
-  env?: Record<string, string>,
-) => {
-  const entries = env ? Object.entries(env) : [];
-  const envArgs =
-    entries.length > 0
-      ? entries
-          .map(
-            ([key, value]) => `export ${key}="${value.replace(/"/g, '\\"')}"`,
-          )
-          .join(" && ") + " && "
-      : "";
-  const ordered: [string, string | undefined][] = [
-    ["pre_install", scripts.pre_install],
-    ["install", scripts.install],
-    ["post_install", scripts.post_install],
-  ];
-  for (const [name, script] of ordered) {
-    if (!script) continue;
-    const target = `/tmp/coder-utils-${name}.sh`;
-    await writeExecutable({
-      containerId: id,
-      filePath: target,
-      content: script,
-    });
-    const resp = await execContainer(id, ["bash", "-c", `${envArgs}${target}`]);
-    if (resp.exitCode !== 0) {
-      console.log(`script ${name} failed:`);
-      console.log(resp.stdout);
-      console.log(resp.stderr);
-      throw new Error(`coder-utils ${name} script exited ${resp.exitCode}`);
-    }
-  }
 };
 
 setDefaultTimeout(60 * 1000);
@@ -260,6 +183,9 @@ describe("codex", async () => {
     await runScripts(id, scripts);
     const resp = await readFileContainer(id, "/home/coder/.codex/config.toml");
     expect(resp).toContain('preferred_auth_method = "apikey"');
+    expect(resp).not.toContain("model_provider");
+    expect(resp).not.toContain("[model_providers.");
+    expect(resp).not.toContain("model_reasoning_effort");
   });
 
   test("pre-post-install-scripts", async () => {
@@ -314,6 +240,21 @@ describe("codex", async () => {
     expect(configToml).toContain('model_provider = "aibridge"');
     expect(configToml).toContain('model_reasoning_effort = "none"');
     expect(configToml).toContain("[model_providers.aibridge]");
+  });
+
+  test("model-reasoning-effort-standalone", async () => {
+    const { id, scripts } = await setup({
+      moduleVariables: {
+        model_reasoning_effort: "high",
+      },
+    });
+    await runScripts(id, scripts);
+    const configToml = await readFileContainer(
+      id,
+      "/home/coder/.codex/config.toml",
+    );
+    expect(configToml).toContain('model_reasoning_effort = "high"');
+    expect(configToml).not.toContain("model_provider");
   });
 
   test("workdir-trusted-project", async () => {
