+++ registry/coder-labs/modules/codex/scripts/install.sh.tftpl
@@ -21,6 +21,8 @@ echo "--------------------------------"
 printf "codex_version: %s\n" "$${ARG_CODEX_VERSION}"
 printf "workdir: %s\n" "$${ARG_WORKDIR}"
 printf "enable_ai_gateway: %s\n" "$${ARG_ENABLE_AI_GATEWAY}"
+printf "install_codex: %s\n" "$${ARG_INSTALL}"
+printf "model_reasoning_effort: %s\n" "$${ARG_MODEL_REASONING_EFFORT}"
 echo "--------------------------------"
 
 function install_codex() {
